Remove not needed check in rounding
Whenever the inputs of LessThan(n) are of bit size bigger than n,
the function might not function properly. The issue is that the
bit decomposition inside LessThan might fail. However, in that
case the constraints in Num2Bits will not be satisfiable, so
there is no need to have additional constraints before calling
Num2Bits. It is enough to assume that den < 2^252.